Migrate some time properties to double

+++ src/ru/nsu/ccfit/zuev/osu/game/GameObjectListener.java
@@ -41,7 +41,7 @@ public interface GameObjectListener {
 
     int getCursorsCount();
 
-    void registerAccuracy(float acc);
+    void registerAccuracy(double acc);
     
     void updateAutoBasedPos(float pX, float pY);
